@@ -43,14 +43,17 @@ public final class KeyConfiguration<K: FXVersioning>: Encodable {
     }
 
     static func resolveKeys(_ keys: [ConfigurationKey], universe: [String]) -> FXSortedSet<String> {
-        return FXSortedSet(keys.reduce(into: [String](), { (result, key) in
-            switch key {
-                case let .literal(val):
-                    result.append(val)
-                case let .prefix(prefix):
-                    result.append(contentsOf: universe.filter { $0.hasPrefix(prefix)} )
-            }
-        }))
+        return FXSortedSet(
+            keys.reduce(
+                into: [String](),
+                { (result, key) in
+                    switch key {
+                    case .literal(let val):
+                        result.append(val)
+                    case .prefix(let prefix):
+                        result.append(contentsOf: universe.filter { $0.hasPrefix(prefix) })
+                    }
+                }))
     }
 
     public func get<T>(_ key: String) -> T? {
@@ -63,17 +66,18 @@ public final class KeyConfiguration<K: FXVersioning>: Encodable {
     public func encode(to encoder: Encoder) throws {
         let allPossibleAllowed = Self.resolveKeys(Array(K.aggregatedConfigurationKeys), universe: Array(self.inputs.keys))
 
-        try encoder.fxEncodeHash(of: try inputs.filter{ (k, _) in allPossibleAllowed.contains(k) }.mapValues { (val: Encodable) -> String in try val.fxEncodeJSON() })
+        try encoder.fxEncodeHash(of: try inputs.filter { (k, _) in allPossibleAllowed.contains(k) }.mapValues { (val: Encodable) -> String in try val.fxEncodeJSON() })
     }
 
     func isNoop() -> Bool {
-        return self.allowedInputs.isEmpty && K.aggregatedConfigurationKeys.allSatisfy({ key in
-            switch key {
+        return self.allowedInputs.isEmpty
+            && K.aggregatedConfigurationKeys.allSatisfy({ key in
+                switch key {
                 case .literal(_):
                     return false
-                case let .prefix(prefix):
+                case .prefix(let prefix):
                     return self.inputs.keys.filter { $0.hasPrefix(prefix) }.isEmpty
-            }
-        })
+                }
+            })
     }
 }
